refactor: Move agent interface into the provider package -- go-llm-router (13)
