+++ cmd/app/buildAgentRegistry.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/pardnchiu/agenvoy/internal/agents/exec"
 	agentKeychain "github.com/pardnchiu/agenvoy/internal/agents/keychain"
-	"github.com/pardnchiu/agenvoy/internal/agents/router"
+	"github.com/pardnchiu/agenvoy/internal/agents/provider/router"
 	agentTypes "github.com/pardnchiu/agenvoy/internal/agents/types"
 	"github.com/pardnchiu/agenvoy/internal/session/config"
 )
